+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- added error tag configuration to enable commuicated processing failues as tag state to paperless instance relevant issue #5 
+
 ### Fixed
 - more robust fetching of custom_field types in case of newly added custom field types